resolve conflicts in constants.go by accepting incoming changes
